requirements.txt
psycopg2
djoser
python-decouple
psycopg2-binary
mysqlclient==2.2.1
anyio==4.1.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.6.0
asttokens==2.4.1
async-lru==2.0.4
attrs==18.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.2.0
ConfigArgParse==1.7
coursera-dl==0.11.5
cryptography==41.0.7
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==5.0.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.19.0
filelock==3.9.0
fonttools==4.47.2
fpdf==1.7.2
fqdn==1.5.1
idna==3.4
importlib-metadata==7.0.0
ipykernel==6.27.1
ipython==8.18.1
isoduration==20.11.0
jaraco.classes==3.3.0
jedi==0.19.1
Jinja2==3.1.3
json2html==1.3.0
json5==0.9.14
jsonpointer==2.4
jupyter_client==8.6.0
jupyter_core==5.7.1
keyring==24.3.0
MarkupSafe==2.1.4
matplotlib-inline==0.1.6
more-itertools==10.2.0
mysqlclient==2.2.1
nest-asyncio==1.6.0
oauthlib==3.2.2
packaging==23.2
parso==0.8.3
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.5.1
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
pyzmq==25.1.2
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==69.0.3
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
tornado==6.4
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
tzdata==2023.4
urllib3==2.1.0
wcwidth==0.2.13
webencodings==0.5.1
zipp==3.17.0
Pillow
